refactor code and add unsupportedResources to getKubeResourcesManifest
