+++ cmd/cronjob_test.go
@@ -12,10 +12,10 @@ func TestCronjobV1(t *testing.T) {
 	file := "../fixtures/cronjob_v1beta1.yml"
 	fileFixed := "../fixtures/cronjob-fixed_v1beta1.yml"
 	assert := assert.New(t)
-	resources, err := getKubeResourcesManifest(file)
+	resources, _, err := getKubeResourcesManifest(file)
 	assert.Nil(err)
 	fixedResources := fix(resources)
-	correctlyFixedResources, err := getKubeResourcesManifest(fileFixed)
+	correctlyFixedResources, _, err := getKubeResourcesManifest(fileFixed)
 	assert.Nil(err)
 	assert.Nil(deep.Equal(correctlyFixedResources, fixedResources))
 }
